hf-spaces-deployment/universal_lora_deploy_2025-10-14/requirements.txt
# SDXL Universal LoRA - FINAL FIXED VERSION - Updated Oct 15 2025
fastapi==0.104.1
uvicorn[standard]==0.24.0
Pillow==10.1.0
pydantic==2.5.0
numpy==1.24.3

# PyTorch - Version ranges to resolve xformers conflict
torch>=2.1.0,<2.2.0
torchvision>=0.16.0,<0.17.0

# Diffusers and transformers - Compatible versions
diffusers==0.24.0
transformers==4.36.0

# LoRA libraries - Version ranges for compatibility
peft>=0.7.0,<0.8.0
safetensors>=0.4.0,<0.5.0
accelerate>=0.25.0,<0.26.0

# Memory optimization - Removed xformers to fix dependency conflict
bitsandbytes>=0.41.0,<0.42.0

# Hugging Face Hub
huggingface-hub>=0.19.0,<0.20.0

# CRITICAL: This fixes the torch==2.1.1 vs xformers==torch.2.1.2 conflict
# Upload THIS file to your HF Space to resolve build errors